:bug: Fix exceptions not shared.
