+++ discordoauth2/__init__.py
@@ -1,5 +1,6 @@
+from .exceptions import exceptions, Exceptions
 from .async_oauth import AsyncClient, AsyncAccessToken, AsyncPartialAccessToken
-from .sync_oauth import Client, AccessToken, PartialAccessToken, exceptions
+from .sync_oauth import Client, AccessToken, PartialAccessToken
 
 __all__ = [
     "Client",
@@ -9,4 +10,5 @@ __all__ = [
     "AsyncAccessToken",
     "AsyncPartialAccessToken",
     "exceptions",
+    "Exceptions",
 ]
