+++ discordoauth2/sync_oauth.py
@@ -1,7 +1,10 @@
 import requests
+
 from datetime import datetime
 from typing import Optional, Union, Literal
 from urllib import parse
+
+from .exceptions import Exceptions
 
 
 class PartialAccessToken:
@@ -23,16 +26,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def fetch_connections(self) -> list[dict]:
         """Retrieves a list of [connection object](https://discord.com/developers/docs/resources/user#connection-object)s the user has linked. Requires the `connections` scope"""
@@ -44,16 +47,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def fetch_guilds(self) -> list[dict]:
         """Retrieves a list of [partial guild](https://discord.com/developers/docs/resources/user#get-current-user-guilds-example-partial-guild)s the user is in. Requires the `guilds` scope"""
@@ -66,16 +69,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def fetch_guild_member(self, guild_id: int) -> dict:
         """Retrieves the user's [guild member object](https://discord.com/developers/docs/resources/guild#guild-member-object) in a specific guild. Requires the `guilds.members.read` scope
@@ -90,18 +93,18 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 404:
-            raise exceptions.HTTPException(f"user is not in this guild.")
+            raise Exceptions.HTTPException(f"user is not in this guild.")
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def join_guild(
         self,
@@ -134,24 +137,24 @@ class PartialAccessToken:
         )
 
         if response.status_code == 204:
-            raise exceptions.HTTPException(f"member is already in the guild.")
+            raise Exceptions.HTTPException(f"member is already in the guild.")
         elif response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 403:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"the Bot token must be for a bot in the guild that has permissions to create invites in the target guild and must have any other required permissions."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def fetch_metadata(self):
         """Retrieves the user's [metadata](https://discord.com/developers/docs/resources/user#application-role-connection-object) for this application. Requires the `role_connections.write` scope"""
@@ -163,16 +166,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def update_metadata(
         self, platform_name: str = None, username: str = None, **metadata
@@ -207,16 +210,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def clear_metadata(self):
         """Clears the user's metadata for this application. Requires the `role_connections.write` scope"""
@@ -229,16 +232,16 @@ class PartialAccessToken:
         if response.ok:
             return response.json()
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
 
 class AccessToken(PartialAccessToken):
@@ -307,16 +310,16 @@ class Client:
         if response.ok:
             return AccessToken(response.json(), self)
         elif response.status_code == 400:
-            raise exceptions.HTTPException(
+            raise Exceptions.HTTPException(
                 "the code, client id, client secret or the redirect uri is invalid/don't match."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def refresh_token(self, refresh_token: str) -> AccessToken:
         """Converts a refresh token into a new `AccessToken`
@@ -336,16 +339,16 @@ class Client:
         if response.ok:
             return AccessToken(response.json(), self)
         elif response.status_code == 400:
-            raise exceptions.HTTPException(
+            raise Exceptions.HTTPException(
                 "the refresh token, client id or client secret is invalid/don't match."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def client_credentails_grant(self, scope: list[str]) -> AccessToken:
         """Creates an `AccessToken` on behalf of the application's owner. If the owner is a team, then only `identify` and `applications.commands.update` are allowed.
@@ -360,16 +363,16 @@ class Client:
         if response.ok:
             return AccessToken(response.json(), self)
         elif response.status_code == 400:
-            raise exceptions.HTTPException(
+            raise Exceptions.HTTPException(
                 "the scope, client id or client secret is invalid/don't match."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def revoke_token(self, token: str, token_type: str = None):
         """Revokes a OAuth2 token related to the client.
@@ -386,16 +389,16 @@ class Client:
         if response.ok:
             return
         elif response.status_code == 401:
-            raise exceptions.Forbidden(
+            raise Exceptions.Forbidden(
                 f"this AccessToken does not have the nessasary scope."
             )
         elif response.status_code == 429:
-            raise exceptions.RateLimited(
+            raise Exceptions.RateLimited(
                 f"You are being Rate Limited. Retry after: {response.json()['retry_after']}",
                 retry_after=response.json()["retry_after"],
             )
         else:
-            raise exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
+            raise Exceptions.HTTPException(f"Unexpected HTTP {response.status_code}")
 
     def generate_uri(
         self,
@@ -429,19 +432,3 @@ class Client:
             "permissions": permissions,
         }
         return f"https://discord.com/oauth2/authorize?{parse.urlencode({key: value for key, value in params.items() if value is not None})}"
-
-
-class exceptions:
-    class BaseException(Exception):
-        pass
-
-    class HTTPException(BaseException):
-        pass
-
-    class RateLimited(HTTPException):
-        def __init__(self, text, retry_after):
-            self.retry_after = retry_after
-            super().__init__(text)
-
-    class Forbidden(HTTPException):
-        pass
